Move filter-tests to new file
